requirements.txt
fastapi
uvicorn
python-multipart
requests
beautifulsoup4
jinja2
openai
pypdf
langchain
langchain-openai
langchain-google-genai
pydantic
selenium
webdriver-manager
python-dotenv
google-generativeai
